+++ surfsense_backend/app/services/docling_service.py
