--- surfsense_backend/app/services/document_processing/__init__.py
@@ -1,1 +1,0 @@
-# Document processing services for SurfSense
